Uploads 2020 livestream artwork & config
Unlike the last two years this actually includes the OBS config files!  Blender source files for the stream background animation are also included, everything else is pretty much the same as previous years.

The background animation video file is much lower quality than the one used in production, I have included it so the config works and can reference the file but if somebody else is to ever use this they should really re-render it in Blender and use the Premiere file alongside the new frames to export out a good quality version.  The frames themselves are about 10GB and I'm not including them either.